requirements.txt
streamlit
pytesseract
Pillow
pandas
requests
PyPDF2
